improvements made to results outputting

+++ modules/cogs/cognitor.py
@@ -12,26 +12,48 @@ class cogs(object):
 	query = {};
 	cogs = {};
 	threads = 4;
-	results = "result.COG.csv"
+	output = "result.COG.csv"
 	software_path = str(Path(__file__).parents[0]) + '/'
 
 	"""docstring for cogs"""
-	def __init__(self):
+	def __init__(self, directory, output = 'result.GOG.csv', reset = 'reset'):
 
 		super(cogs, self).__init__()
+		self.output = output
+		self.directory = directory
+		self._createDirectory(reset);
 
-	def _createDirectory(self):
+	def _createDirectory(self, reset = 'reset'):
 		
-		if os.path.exists(self.directory):
-			shutil.rmtree(self.directory)
+		if reset == 'reset' :
 
-		if not os.path.exists(self.directory):
-		    os.makedirs(self.directory)
-		    os.makedirs(self.directory + "BLASTss")
-		    os.makedirs(self.directory + "BLASTno")
-		    os.makedirs(self.directory + "BLASTff")
-		    os.makedirs(self.directory + "BLASTcogn")
-		    os.makedirs(self.directory + "BLASTconv")
+			try:
+
+				if os.path.exists(self.directory):
+					shutil.rmtree(self.directory)
+
+				pass
+
+			except Exception as e:
+
+				pass
+
+		try:
+
+			if not os.path.exists(self.directory):
+			    os.makedirs(self.directory)
+			    os.makedirs(self.directory + "BLASTss")
+			    os.makedirs(self.directory + "BLASTno")
+			    os.makedirs(self.directory + "BLASTff")
+			    os.makedirs(self.directory + "BLASTcogn")
+			    os.makedirs(self.directory + "BLASTconv")
+
+			pass
+
+		except Exception as e:
+
+			print(str(e))
+			os._exit(1)
 
 	def _fixFiles(self):
 
@@ -65,11 +87,11 @@ class cogs(object):
 
 	def _queryDb(self):
 		
-		os.system("makeblastdb -in " + self.query['sequences'] + " -dbtype prot -out " + self.query['db'] );
+		os.system("makeblastdb -in " + self.query['sequences'] + " -dbtype prot -logfile " + self.directory + "queryBlastLog.log -out " + self.query['db'] );
 
 
 	def _cogsBlast(self):
-		
+	
 		os.system("psiblast -query " + self.query['sequences'] + " -db " + self.query['db'] + " -show_gis -outfmt 7 -num_threads "+ str(self.threads) +" -num_alignments 10 -dbsize 100000000 -comp_based_stats F -seg no -out " + self.directory + "BLASTss/QuerySelf.tab" );
 		os.system("psiblast -query " + self.query['sequences'] + " -db " + self.cogs['db'] + " -show_gis -outfmt 7 -num_threads "+ str(self.threads) +" -num_alignments 1000 -dbsize 100000000 -comp_based_stats F -seg no -out " + self.directory + "BLASTno/QueryCOGs.tab" );
 		os.system("psiblast -query " + self.query['sequences'] + " -db " + self.cogs['db'] + " -show_gis -outfmt 7 -num_threads "+ str(self.threads) +" -num_alignments 1000 -dbsize 100000000 -comp_based_stats T -seg yes -out " + self.directory + "BLASTff/QueryCOGs.tab" );
@@ -80,7 +102,7 @@ class cogs(object):
 
 		os.system(self.software_path + "COG_software/COGmakehash/COGmakehash -i=" + self.directory + "tmp.p2o.csv -o=" + self.directory + "BLASTcogn -s=',' -n=1" );
 		os.system(self.software_path + "COG_software/COGreadblast/COGreadblast -d=" + self.directory + "BLASTcogn -u=" + self.directory + "BLASTno -f=" + self.directory + "BLASTff -s=" + self.directory + "BLASTss -e=0.1 -q=2 -t=2" );
-		os.system(self.software_path + "COG_software/COGcognitor/COGcognitor -i=" + self.directory + "BLASTcogn -t=" + self.cogs['csv'] + " -q=" + self.query['csv'] + " -o=" + self.results  );
+		os.system(self.software_path + "COG_software/COGcognitor/COGcognitor -i=" + self.directory + "BLASTcogn -t=" + self.cogs['csv'] + " -q=" + self.query['csv'] + " -o=" + self.output  );
 
 	def setQuery(self, sequences):
